requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
beautifulsoup4==4.11.1
    # via django-bootstrap-v5
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
coverage==7.0.5
    # via -r requirements.in
cryptography==39.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==1.2.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==4.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-bootstrap-v5
    #   django-debug-toolbar
    #   django-htmx
    #   django-modeltranslation
django-allauth==0.52.0
    # via -r requirements.in
django-bootstrap-v5==1.0.11
    # via -r requirements.in
django-cache-url==3.4.4
    # via environs
django-debug-toolbar==3.8.1
    # via -r requirements.in
django-filebrowser==4.0.2
    # via -r requirements.in
django-grappelli==3.0.4
    # via
    #   -r requirements.in
    #   django-filebrowser
django-htmx==1.13.0
    # via -r requirements.in
django-modeltranslation==0.18.7
    # via -r requirements.in
django-tinymce==3.5.0
    # via -r requirements.in
environs[django]==9.5.0
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==16.4.0
    # via factory-boy
idna==3.4
    # via requests
marshmallow==3.19.0
    # via environs
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.0
    # via marshmallow
pillow==9.4.0
    # via django-filebrowser
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
python-dateutil==2.8.2
    # via faker
python-dotenv==0.21.0
    # via environs
python3-openid==3.2.0
    # via django-allauth
requests==2.28.2
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   django-filebrowser
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.4.0
    # via django-modeltranslation
tzdata==2022.7
    # via django
urllib3==1.26.14
    # via requests
